Move focus analysis to it's own class. Start moing autofocus runner to it's own class
